Refactor to replace PocketBase with a new data module across components
